+++ src/configuration/validate_configuration.ts
@@ -5,11 +5,6 @@ export function validateConfiguration(
   configuration: IConfiguration,
   logger: ILogger
 ): void {
-  if (configuration.publishQuiet) {
-    logger.warn(
-      '`publishQuiet` option is no longer needed, you can remove it from your configuration; see https://github.com/cucumber/cucumber-js/blob/main/docs/deprecations.md'
-    )
-  }
   if (configuration.requireModule.length && !configuration.require.length) {
     logger.warn(
       'Use of `require-module` option normally means you should specify your support code paths with `require`; see https://github.com/cucumber/cucumber-js/blob/main/docs/configuration.md#finding-your-code'